% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame.ms2spectra}
\alias{as.data.frame.ms2spectra}
\title{Tidy-Quant data object print definition}
\usage{
\method{as.data.frame}{ms2spectra}(x, ...)
}
\arguments{
\item{x}{rmfasta data object}

\item{...}{unused legacy}
}
\value{
print object summary
}
\description{
Tidy-Quant data object print definition
}
