% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_object.R
\name{check_msNspectra}
\alias{check_msNspectra}
\title{Check the integrity of a msNspectra data object}
\usage{
check_msNspectra(x = NULL)
}
\arguments{
\item{x}{An msNspectra data object}
}
\value{
silent on success, an abort message on fail
}
\description{
\code{check_msNspectra()} is a helper function that checks the structure and contents of
an msNspectra data object
}
