% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_psms_comet.R
\name{import_comet}
\alias{import_comet}
\title{Read a Comet proteomic search results file}
\usage{
import_comet(path, cpus = 1)
}
\arguments{
\item{path}{String path to file for importing}

\item{cpus}{The number of cpus to use for importing}
}
\value{
a tibble
}
\description{
\code{import_comet()} is the helper function to import tandem ms search results
into a standardized data table.
}
