% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ms2_mfg.R
\name{import_mgf}
\alias{import_mgf}
\title{The main function for parsing a fasta file}
\usage{
import_mgf(path = NULL)
}
\arguments{
\item{path}{a character string of the path to the MGF formatted file}
}
\value{
a tibble
}
\description{
\code{import_mgf()} get the current regex
}
