% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ms2_mzml.R
\name{import_mzml}
\alias{import_mzml}
\title{The main function for parsing a fasta file}
\usage{
import_mzml(path = NULL)
}
\arguments{
\item{path}{a character string of the path to the MGF formatted file}
}
\value{
a tibble
}
\description{
\code{import_mzml()} get the current regex
}
