% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_psms_sage.R
\name{import_sage}
\alias{import_sage}
\title{Read a Sage proteomic search results file}
\usage{
import_sage(path)
}
\arguments{
\item{path}{String path to file for importing}
}
\value{
a tibble
}
\description{
\code{import_sage()} is the helper function to import tandem ms search results
into a standardized data table.
}
