% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_psms_tide.R
\name{import_tide}
\alias{import_tide}
\title{Read a OMSSA proteomic search results file}
\usage{
import_tide(path)
}
\arguments{
\item{path}{String path to file for importing}
}
\value{
a tibble
}
\description{
\code{import_tide()} is the helper function to import tandem ms search results
into a standardized data table.
}
