% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_psms_xtandem.R
\name{import_xtandem}
\alias{import_xtandem}
\title{helper function to read in platform specific results}
\usage{
import_xtandem(x, cpus = 1)
}
\arguments{
\item{x}{location of file to parse}
}
\value{
a tibble
}
\description{
helper function to read in platform specific results
}
