% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_object.R
\name{ms2spectra}
\alias{ms2spectra}
\title{Declaring the spectra data object}
\usage{
ms2spectra(obj)
}
\arguments{
\item{obj}{The current tibble or list object}
}
\description{
Declaring the spectra data object
}
