% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_psms_msamanda.R
\name{ms_amanda_peptide}
\alias{ms_amanda_peptide}
\title{Format MS Amanda peptide strings}
\usage{
ms_amanda_peptide(sequence = NULL, modifications = NULL)
}
\arguments{
\item{sequence}{A peptide string}

\item{modifications}{A string of modifications}
}
\value{
string
}
\description{
Format MS Amanda peptide strings
}
