% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_object.R
\name{msNspectra}
\alias{msNspectra}
\title{Declaring the spectra data object}
\usage{
msNspectra(obj)
}
\arguments{
\item{obj}{The current tibble or list object}
}
\description{
Declaring the spectra data object
}
