% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_subset.R
\name{msreadr_quo_name}
\alias{msreadr_quo_name}
\title{Helper function to get a name from the ...}
\usage{
msreadr_quo_name(...)
}
\arguments{
\item{...}{a quo}
}
\value{
a character string
}
\description{
Helper function to get a name from the ...
}
