% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_psms_omssa.R
\name{omssa_peptide}
\alias{omssa_peptide}
\title{Format OMSSA peptide strings}
\usage{
omssa_peptide(sequence = NULL, modifications = NULL)
}
\arguments{
\item{sequence}{A peptide string}

\item{modifications}{A string of modifications}
}
\value{
string
}
\description{
Format OMSSA peptide strings
}
