% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{plot.ms2spectra}
\alias{plot.ms2spectra}
\title{A helper function for the print definition of a ms2spectra object}
\usage{
\method{plot}{ms2spectra}(x = NULL, ..., mz = NULL, mz_tolerance = 0.1)
}
\arguments{
\item{x}{An ms2spectra data object}

\item{...}{Unused legacy}

\item{mz}{The mz value (Th) to extract}

\item{mz_tolerance}{The tolerance +/- the mz value (Th) to extract}
}
\description{
A helper function for the print definition of a ms2spectra object
}
