% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{plot.msNspectra}
\alias{plot.msNspectra}
\title{A helper function for the print definition of a msNspectra object}
\usage{
\method{plot}{msNspectra}(
  x = NULL,
  ...,
  type = c("IC", "3D"),
  mz = NULL,
  mz_tolerance = 0.1,
  bins = 32
)
}
\arguments{
\item{x}{An msNspectra data object}

\item{...}{Unused legacy}

\item{mz}{The mz value (Th) to extract}

\item{mz_tolerance}{The tolerance +/- the mz value (Th) to extract}
}
\description{
A helper function for the print definition of a msNspectra object
}
