% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_print.R
\name{print.ms2spectra}
\alias{print.ms2spectra}
\title{A helper function for the print definition of a ms2spectra object}
\usage{
\method{print}{ms2spectra}(x, ...)
}
\arguments{
\item{x}{An ms2spectra data object}

\item{...}{Unused legacy}
}
\description{
A helper function for the print definition of a ms2spectra object
}
