% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_print.R
\name{print.msNspectra}
\alias{print.msNspectra}
\title{A helper function for the print definition of a msNspectra object}
\usage{
\method{print}{msNspectra}(x, ...)
}
\arguments{
\item{x}{An msNspectra data object}

\item{...}{Unused legacy}
}
\description{
A helper function for the print definition of a msNspectra object
}
