% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_psms.R
\name{read_psms}
\alias{read_psms}
\title{Read a proteomic search results file from Comet, MS Amanda, OMSSA, Sage and X!Tandem}
\usage{
read_psms(
  path_to_psms = NULL,
  path_to_mzml = NULL,
  platform = c("comet", "ms_amanda", "omssa", "sage", "xtandem", "tide"),
  ...
)
}
\arguments{
\item{path_to_psms}{String path to file for importing}

\item{path_to_mzml}{The number of cpus to use for importing}

\item{platform}{The number of cpus to use for importing}

\item{...}{pass-through arguments}
}
\value{
a tibble
}
\description{
\code{read_psms()} is the main function for importing tandem ms search results
into a standardized data table from various platforms.
}
