% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectra.R
\name{read_spectra}
\alias{read_spectra}
\title{Read and ms2 spectra file}
\usage{
read_spectra(path = NULL, include_spectra = TRUE)
}
\arguments{
\item{path}{A character string of the path to the mzML or MGF formatted file}

\item{include_spectra}{A boolean to keep/drop the nested MSn spectrum}
}
\description{
\code{read_spectra()} is the main function for parse a given mzML or MGF file into
a ms2spectra standardized data object for use among the tidyproteomics packages.
}
