% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_subset.R
\name{subset}
\alias{subset}
\title{Create a data subset}
\usage{
subset(data = NULL, ..., .verbose = TRUE)
}
\arguments{
\item{data}{An ms2spectra data object}

\item{...}{A three part expression (eg. x == a)}

\item{.verbose}{A boolean to print messages}
}
\description{
\code{subset()} is the main function for sub-setting spectra data from a ms2spectra
data-object based on a regular expression and targeted annotation. This function
will return a smaller ms2spectra data-object.
}
\examples{
# creates a subset of just of spectra
data <- path_to_example() |> read_spectra()

data |> subset(ms_event == 3)

}
