% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_subset.R
\name{subset_helper}
\alias{subset_helper}
\title{Helper function to do the subsetting ...}
\usage{
subset_helper(tbl = NULL, str_quo = NULL, .verbose = TRUE)
}
\arguments{
\item{tbl}{... the data object}

\item{str_quo}{... a quo}

\item{.verbose}{... T/F to print progress}
}
\value{
a character string
}
\description{
Helper function to do the subsetting ...
}
