% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mfg.R
\name{write_mgf}
\alias{write_mgf}
\title{Write an MGF spectrum file}
\usage{
write_mgf(data = NULL, path = NULL)
}
\arguments{
\item{data}{An ms2spectra data object}

\item{path}{A character string of the path to the MGF formatted file}
}
\value{
silent on success, an abort message on fail
}
\description{
\code{write_mgf()} write the contents of an ms2spectrum data object to an MGF
formmated file.
}
\examples{
\dontrun{
data <- path_to_example() |> read_spectra()

data |> write_mgf("path_to_file.mgf")
}

}
