% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mzml.R
\name{write_mzml}
\alias{write_mzml}
\title{Write an mzML spectrum file}
\usage{
write_mzml(data = NULL, path = NULL)
}
\arguments{
\item{data}{An ms2spectra data object}

\item{path}{A character string of the path to the mzML formatted file}
}
\value{
silent on success, an abort message on fail
}
\description{
\code{write_mzml()} write the contents of an ms2spectrum data object to an mzML
formmated file.
}
\examples{
\dontrun{
data <- path_to_example() |> read_spectra()

data |> write_mzml("path_to_file.mzML")
}
}
