% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_xic.R
\name{xic}
\alias{xic}
\title{Create a data subset}
\usage{
xic(data = NULL, mz = NULL, mz_tolerance = 0.1)
}
\arguments{
\item{data}{An ms2spectra data object}

\item{mz}{The mz value (Th) to extract}

\item{mz_tolerance}{The tolerance +/- the mz value (Th) to extract}
}
\description{
\code{subset()} is the main function for sub-setting spectra data from a ms2spectra
data-object based on a regular expression and targeted annotation. This function
will return a smaller ms2spectra data-object.
}
