% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame.ms2spectra}
\alias{as.data.frame.ms2spectra}
\title{Help function to convert a spectra data object to a data.frame}
\usage{
\method{as.data.frame}{ms2spectra}(x, ...)
}
\arguments{
\item{x}{The spectra data object}

\item{...}{Unused legacy}
}
\description{
Help function to convert a spectra data object to a data.frame
}
