% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ms2_mfg.R
\name{import_mgf}
\alias{import_mgf}
\title{Parse an MGF file}
\usage{
import_mgf(path = NULL)
}
\arguments{
\item{path}{A character string of the path to the MGF formatted file}
}
\description{
\code{import_mgf()} a helper function to parse a given MGF file into a ms2spectra
standardized data object for use among the tidyproteomics packages.
}
