% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_psms_msamanda.R
\name{import_msamanda}
\alias{import_msamanda}
\title{Read a MS Amanda proteomic search results file}
\usage{
import_msamanda(path, cpus = 1)
}
\arguments{
\item{path}{String path to file for importing}

\item{cpus}{The number of cpus to use for importing}
}
\value{
a tibble
}
\description{
\code{import_msamanda()} is the helper function to import tandem ms search results
into a standardized data table.
}
