% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_psms_omssa.R
\name{import_omssa}
\alias{import_omssa}
\title{Read a OMSSA proteomic search results file}
\usage{
import_omssa(path, cpus = 1)
}
\arguments{
\item{path}{String path to file for importing}

\item{cpus}{The number of cpus to use for importing}
}
\value{
a tibble
}
\description{
\code{import_omssa()} is the helper function to import tandem ms search results
into a standardized data table.
}
