% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_psms_xtandem.R
\name{import_xtandem}
\alias{import_xtandem}
\title{Read a X!Tandem proteomic search results file}
\usage{
import_xtandem(path)
}
\arguments{
\item{path}{String path to file for importing}
}
\value{
a tibble
}
\description{
\code{import_xtandem()} is the helper function to import tandem ms search results
into a standardized data table.
}
