% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mzml.R
\name{write_mzml}
\alias{write_mzml}
\title{The main function for parsing a fasta file}
\usage{
write_mzml(data = NULL, path = NULL)
}
\arguments{
\item{path}{a character string of the path to the MGF formatted file}
}
\value{
a tibble
}
\description{
\code{import_mgf()} get the current regex
}
